Sunny Day and Rain Dance replace Splash and Tail Whip
Not completed, still need the accuracy boost on Thunder and charge effect change on solar beam, but does boost water, fire, grass and electric respectively. Also, may continue for a turn or two after the battle ends and start a new battle, but I don't think of that as a serious bug.