make wobbuffet ai not totally idiotic
it was spamming counter or mirror coat based on 'type effectiveness'  which is the exact opposite of how you want these used, you want to use them more randomly so the player can't predict. Tried to prevent spam based on type effectiveness in general for 1 power moves but not sure that worked.