minor messing with town maps and music for possible orre and johto later
the music here hasn't actually been added, nor the extra maps, just placed in the folders; I have not tested the music asm to see if it actually works in game but it looks like it would be easy to modify if needed, if perhaps a bit of a pain in the rear to go line by line. I used  tilemap studio to edit the town map rle files. As always, feel free to use my work.