broken color, rom won't load
Tried a straight copy over of code from red color. Had quite a few issues, but decided to comment in the heavily broken code just in case someone wants to look and see if they can figure out everything that went horribly wrong. I suspect a lot of changes were unneeded and just some palette loading mech is needed on maps specifically, but the need to load for each sprite puzzles me a bit and wonder if the rst vectors are actually needed or not...