bunch of tiny Johto map changes plus altered forest tileset
The cut down tree is now a proper tree, so that I can make a better looking version of Johto Ilex Forest map. I now have most of the Johto outdoor maps layouts but obviously not the scripts. I plan to re-use existing Pokecenter maps and marts maps so as to take up less space if possible.